move gwdi to processing subfolder
