+++ src/toolbox_continu_inzicht/gwdi/processing/__init__.py
@@ -1,0 +1,7 @@
+from toolbox_continu_inzicht.gwdi.processing.gwdi_inference import GwdiInference
+from toolbox_continu_inzicht.gwdi.processing.gwdi_model import GwdiModel
+
+__all__ = [
+    "GwdiInference",
+    "GwdiModel",
+]
